refactored RemoteitApiSession
