+++ Remoteit.Test/RestApi/RemoteitApiSessionTest.cs
@@ -1,0 +1,20 @@
+﻿using Xunit;
+using Remoteit.RestApi;
+
+namespace Remoteit.Test.RestApi
+{
+    public class RemoteitApiSessionTest
+    {
+        [Fact]
+        public void TestSessionExpiredCalculation()
+        {
+            bool hasExpired = true;
+        }
+
+        [Fact]
+        public void TestSessionNotxpiredCalculation()
+        {
+            bool hasExpired = true;
+        }
+    }
+}
